there is no Parser class
adding it would make code more OOP

Refactor code to add Parser class
