+++ src/main/java/DialogBox.java
@@ -1,3 +1,6 @@
+import java.io.IOException;
+import java.util.Collections;
+
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
@@ -9,8 +12,6 @@ import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
 
-import java.io.IOException;
-import java.util.Collections;
 
 /**
  * Represents a dialog box consisting of an ImageView to represent the speaker's face
@@ -36,7 +37,6 @@ public class DialogBox extends HBox {
         displayPicture.setImage(img);
         this.getStyleClass().add("dialog-box");
         this.getStyleClass().add(styleClass);
-
 
 
     }
