--- src/main/java/Duke.java
@@ -1,12 +1,0 @@
-public class Duke {
-    public static void main(String[] args) {
-        System.out.println("Hello!");
-    }
-
-    /**
-     * Generates a response for the user's chat message.
-     */
-    public String getResponse(String input) {
-        return "Duke heard: " + input;
-    }
-}
